+++ src/OpenTelemetry/Metrics/Reader/MetricReaderExt.cs
@@ -18,6 +18,7 @@ public abstract partial class MetricReader
     private readonly Lock instrumentCreationLock = new();
     private int metricLimit;
     private int cardinalityLimit;
+    private bool enableLazyAllocation;
     private Metric?[] metrics = [];
     private Metric[] metricsCurrentBatch = [];
     private int metricIndex = -1;
@@ -82,7 +83,8 @@ public abstract partial class MetricReader
                         metricStreamIdentity,
                         this.GetAggregationTemporality(metricStreamIdentity.InstrumentType),
                         this.cardinalityLimit,
-                        exemplarFilter);
+                        exemplarFilter,
+                        enableLazyAllocation: this.enableLazyAllocation);
                 }
                 catch (NotSupportedException nse)
                 {
@@ -170,7 +172,8 @@ public abstract partial class MetricReader
                         this.GetAggregationTemporality(metricStreamIdentity.InstrumentType),
                         metricStreamConfig?.CardinalityLimit ?? this.cardinalityLimit,
                         exemplarFilter,
-                        metricStreamConfig?.ExemplarReservoirFactory);
+                        metricStreamConfig?.ExemplarReservoirFactory,
+                        this.enableLazyAllocation);
 
                     this.instrumentIdentityToMetric[metricStreamIdentity] = metric;
                     this.metrics[index] = metric;
@@ -187,6 +190,7 @@ public abstract partial class MetricReader
     internal void ApplyParentProviderSettings(
         int metricLimit,
         int cardinalityLimit,
+        bool enableLazyAllocation,
         ExemplarFilterType? exemplarFilter,
         ExemplarFilterType? exemplarFilterForHistograms)
     {
@@ -194,6 +198,7 @@ public abstract partial class MetricReader
         this.metrics = new Metric[metricLimit];
         this.metricsCurrentBatch = new Metric[metricLimit];
         this.cardinalityLimit = cardinalityLimit;
+        this.enableLazyAllocation = enableLazyAllocation;
         this.exemplarFilter = exemplarFilter;
         this.exemplarFilterForHistograms = exemplarFilterForHistograms;
     }